pyproject.toml
[tool.poetry]
name = "traycortex"
version = "0.1.0"
description = "borgmatic systray icon"
authors = ["Sebastian Stark <sstark@mailbox.org>"]
license = "ISC"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pystray = "^0.19.5"
pillow = "^10.3.0"
pygobject = "^3.48.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
traycortex = "traycortex.tray:app"
